refactor: move compiler to own package
