+++ packages/api/compiler.js
@@ -1,9 +1,5 @@
 const path = require("path");
-const loadConfig = require("@patternplate/load-config");
-const webpackEntry = require("@patternplate/webpack-entry");
-const MemoryFS = require("memory-fs");
-const resolveFrom = require("resolve-from");
-const webpack = require("webpack");
+const compiler = require("@patternplate/compiler");
 const Observable = require("zen-observable");
 
 module.exports = createCompiler;
@@ -11,68 +7,20 @@ module.exports = createCompiler;
 const debug = require("util").debuglog("PATTERNPLATE");
 
 async function createCompiler({ cwd, target = "" }) {
-  const fs = new MemoryFS();
-  const { config, filepath } = await loadConfig({ cwd });
-
-  const components = await webpackEntry(config.entry, { cwd });
-  const entry = { components };
-
-  if (target === "node") {
-    entry.render = await getEntry(config.render, { filepath });
-  }
-
-  if (target === "web") {
-    entry.mount = await getEntry(config.mount, { filepath });
-    entry.demo = require.resolve("@patternplate/demo-client");
-    entry.probe = require.resolve("@patternplate/probe-client")
-  }
-
-  const compiler = webpack({
-    entry,
-    target,
-    module: {
-      rules: [
-        {
-          test: /\.css$/,
-          use: ["to-string-loader", "css-loader"]
-        },
-        {
-          test: /\.html$/,
-          use: ["html-loader"]
-        }
-      ]
-    },
-    output: {
-      library: "patternplate-[name]",
-      libraryTarget: target === "node" ? "commonjs2" : "window",
-      path: "/",
-      filename: `patternplate.${target}.[name].js`
-    },
-    plugins:
-      target === "web"
-        ? [
-            new webpack.optimize.CommonsChunkPlugin({
-              name: "vendors",
-              minChunks: mod =>
-                mod.context && mod.context.indexOf("node_modules") > -1
-            })
-          ]
-        : []
-  });
-
-  compiler.outputFileSystem = fs;
+  const c = await compiler({cwd, target});
+  const fs = c.outputFileSystem;
 
   const queue = [];
   let listeners = [];
   const next = (message) => listeners.forEach(listener => listener.next(message));
 
-  compiler.plugin("compile", () => {
+  c.plugin("compile", () => {
     debug("compile", target);
     queue.unshift({type: 'start', target, payload: {}});
     next(queue);
   });
 
-  compiler.plugin("done", (stats) => {
+  c.plugin("done", (stats) => {
     if (stats.compilation.errors && stats.compilation.errors.length > 0) {
       debug("error", target);
       queue.unshift({type: 'error', target, payload: stats.compilation.errors});
@@ -83,7 +31,7 @@ async function createCompiler({ cwd, target = "" }) {
     next(queue);
   });
 
-  compiler.plugin("failed", err => {
+  c.plugin("failed", err => {
     debug("failed", target);
     queue.unshift({type: 'error', target, payload: err});
     next(queue);
@@ -94,7 +42,7 @@ async function createCompiler({ cwd, target = "" }) {
   const observable = new Observable(observer => {
     if (!watching) {
       watching = true;
-      compiler.watch({ignored: "**/pattern.json"}, () => {});
+      c.watch({ignored: "**/pattern.json"}, () => {});
     }
 
     listeners.push(observer);
@@ -103,13 +51,8 @@ async function createCompiler({ cwd, target = "" }) {
     }
   });
 
-  observable.compiler = compiler;
+  observable.compiler = c;
   observable.queue = queue;
 
   return observable;
-}
-
-function getEntry(id, { filepath }) {
-  const base = filepath ? path.dirname(filepath) : process.cwd();
-  return resolveFrom(base, id);
 }
