+++ packages/api/package.json
@@ -25,12 +25,11 @@
   },
   "dependencies": {
     "@marionebl/sander": "^0.6.1",
+    "@patternplate/compiler": "^2.0.0-16",
     "@patternplate/load-config": "^2.0.0-16",
     "@patternplate/load-docs": "^2.0.0-16",
     "@patternplate/load-meta": "^2.0.0-16",
     "@patternplate/pack": "^2.0.0-16",
-    "@patternplate/probe-client": "^2.0.0-16",
-    "@patternplate/webpack-entry": "^2.0.0-16",
     "aggregate-error": "^1.0.0",
     "arson": "^0.2.6",
     "chokidar": "^1.7.0",
